Move SCA tests into a new package name
