+++ src/test/java/com/checkmarx/flow/cucumber/integration/sca_scanner/RunSCARemoteRepoSteps.java
@@ -1,4 +1,4 @@
-package com.checkmarx.flow.cucumber.integration.sca;
+package com.checkmarx.flow.cucumber.integration.sca_scanner;
 
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
@@ -7,7 +7,7 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(plugin = { "pretty", "summary", "html:build/cucumber/integration/sca", "json:build/cucumber/integration/sca/cucumber.json" },
         features = "src/test/resources/cucumber/features/integrationTests/sca/scanResultsProcessing.feature",
-        glue = {"com.checkmarx.flow.cucumber.integration.sca"},
+        glue = {"com.checkmarx.flow.cucumber.integration.sca_scanner"},
         tags = "@ScaIntegrationTests and not @Skip")
 public class RunSCARemoteRepoSteps {
 }
